implemented catridge for gameplay
